fl-ocr-system/requirements.txt
flwr>=1.5.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
peft>=0.6.0
Pillow>=10.0.0
jiwer>=3.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.15.0
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.65.0
